pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Project for workshop build your own chatbot."
authors = [
    { name = "Martin Kovacs", email = "martin-kovacs@web.de" }
]
dependencies = [
    "langchain>=0.2.6",
    "langchain-openai>=0.1.14",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.3.4",
    "python-dotenv>=1.0.1",
    "openai>=1.35.10",
    "ruamel-yaml>=0.18.6",
    "jupyter>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[project.scripts]
"example-project-with-rye" = "example_project_with_rye:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
universal = true
#virtual = true # for docker, see https://rye.astral.sh/guide/docker/
dev-dependencies = [
    "pre-commit>=3.7.1",
    "pytest>=8.2.2",
    "pyright>=1.1.370",
]


[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/example_project_with_rye"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 160  # widescreen monitor support :)
target-version = "py311"
# can be upgraded to 3.12 for e.g. nested quotes in f-strings.
# That requires all team members to use Python 3.12.

[tool.pytest.ini_options]
# https://pytest.org/en/7.3.x/explanation/goodpractices.html#which-import-mode
addopts = ["--import-mode=importlib", ]
pythonpath = "src"
